requirements.txt
langchain
PyPDF2
openai
langchain_openai
python-dotenv